services/adapter-evenements/requirements.txt
kafka-python-ng==2.2.3
pydantic

